cloudinary delete url

+++ app/api/blog/route.ts
@@ -48,7 +48,6 @@ export async function GET(req: NextRequest) {
         return NextResponse.json({ error: "Blog not found" }, { status: 404 });
       }
 
-      // Format updatedAt
       const formattedBlog = {
         ...blog,
         updatedAt: formatDate(blog.updatedAt.toISOString()),
@@ -67,7 +66,6 @@ export async function GET(req: NextRequest) {
       const nextCursor = hasNextPage ? blogs[blogs.length - 1].id : null;
       const trimmedBlogs = hasNextPage ? blogs.slice(0, -1) : blogs;
 
-      // Format updatedAt for all blogs
       const formattedBlogs = trimmedBlogs.map((blog) => ({
         ...blog,
         updatedAt: formatDate(blog.updatedAt.toISOString()),
@@ -200,8 +198,16 @@ export async function PUT(req: NextRequest) {
 
     let imageUrl = existingBlog.coverImage;
 
-    const isBase64 = coverImage.startsWith("data:image/");
-    if (isBase64) {
+    if (
+      coverImage !== existingBlog.coverImage &&
+      coverImage.startsWith("data:image/")
+    ) {
+
+      const publicId = existingBlog.coverImage.split("/").pop()?.split(".")[0];
+      if (publicId) {
+        await cloudinary.uploader.destroy(`blogs/${publicId}`);
+      }
+
       const cloudinaryResponse = await cloudinary.uploader.upload(coverImage, {
         folder: "blogs",
         public_id: title.toLowerCase().replace(/\s+/g, "-"),
@@ -263,6 +269,19 @@ export async function DELETE(req: NextRequest) {
   try {
     if (!slug) {
       return NextResponse.json({ error: "Slug is required" }, { status: 400 });
+    }
+
+    const blog = await prisma.blog.findUnique({
+      where: { slug },
+    });
+
+    if (!blog) {
+      return NextResponse.json({ error: "Blog not found" }, { status: 404 });
+    }
+
+    const publicId = blog.coverImage.split("/").pop()?.split(".")[0];
+    if (publicId) {
+      await cloudinary.uploader.destroy(`blogs/${publicId}`);
     }
 
     await prisma.blog.delete({
